Ref : 즐겨찾기 관련 API 분리
